move luigiContext and isInIframe to own composables
