+++ frontend/src/components/dialogs/GProjectDialog.vue
@@ -105,7 +105,6 @@ import {
 } from '@vuelidate/validators'
 import { useRouter } from 'vue-router'
 
-import { useAppStore } from '@/store/app'
 import { useConfigStore } from '@/store/config'
 import { useProjectStore } from '@/store/project'
 
@@ -115,6 +114,7 @@ import GProjectCostObject from '@/components/GProjectCostObject.vue'
 
 import { useLogger } from '@/composables/useLogger'
 import { useProvideProjectContext } from '@/composables/useProjectContext'
+import { useLuigiContext } from '@/composables/useLuigiContext'
 
 import {
   messageFromErrors,
@@ -146,7 +146,7 @@ const emit = defineEmits([
 ])
 
 const logger = useLogger()
-const appStore = useAppStore()
+const luigiContext = useLuigiContext()
 const configStore = useConfigStore()
 const projectStore = useProjectStore()
 const router = useRouter()
@@ -157,7 +157,7 @@ const {
   description,
   purpose,
 } = useProvideProjectContext({
-  appStore,
+  luigiContext,
   configStore,
 })
