+++ frontend/src/composables/useProjectContext.js
@@ -12,14 +12,13 @@ import {
   provide,
 } from 'vue'
 
-import { useAppStore } from '@/store/app'
 import { useConfigStore } from '@/store/config'
 
 import { cleanup } from '@/composables/helper'
-
-import { useProjectShootCustomFields } from './useProjectShootCustomFields'
-import { useProjectMetadata } from './useProjectMetadata'
-import { useProjectCostObject } from './useProjectCostObject'
+import { useLuigiContext } from '@/composables/useLuigiContext'
+import { useProjectShootCustomFields } from '@/composables/useProjectShootCustomFields'
+import { useProjectMetadata } from '@/composables/useProjectMetadata'
+import { useProjectCostObject } from '@/composables/useProjectCostObject'
 
 import cloneDeep from 'lodash/cloneDeep'
 import get from 'lodash/get'
@@ -28,7 +27,7 @@ import set from 'lodash/set'
 
 export function createProjectContextComposable (options = {}) {
   const {
-    appStore = useAppStore(),
+    luigiContext = useLuigiContext(),
     configStore = useConfigStore(),
   } = options
 
@@ -63,9 +62,9 @@ export function createProjectContextComposable (options = {}) {
 
   function createProjectManifest () {
     manifest.value = {}
-    if (appStore.accountId) {
+    if (luigiContext.accountId) {
       set(manifest.value, ['metadata', 'label', 'openmfp.org/managed-by'], 'true')
-      set(manifest.value, ['metadata', 'annotations', 'openmfp.org/account-id'], appStore.accountId)
+      set(manifest.value, ['metadata', 'annotations', 'openmfp.org/account-id'], luigiContext.accountId)
     }
     initialManifest.value = cloneDeep(normalizedManifest.value)
   }
