+++ frontend/src/store/app.js
@@ -8,16 +8,8 @@ import {
   defineStore,
   acceptHMRUpdate,
 } from 'pinia'
-import {
-  ref,
-  watch,
-  toRef,
-  computed,
-} from 'vue'
-import LuigiClient from '@luigi-project/client'
+import { ref } from 'vue'
 import { useNotification } from '@kyvg/vue3-notification'
-
-import { useLogger } from '@/composables/useLogger'
 
 import { parseWarningHeader } from '@/utils/headerWarnings'
 import { errorDetailsFromError } from '@/utils/error'
@@ -26,8 +18,6 @@ import moment from '@/utils/moment'
 import assign from 'lodash/assign'
 
 export const useAppStore = defineStore('app', () => {
-  const logger = useLogger()
-
   const ready = ref(false)
   const sidebar = ref(true)
   const redirectPath = ref(null)
@@ -38,62 +28,8 @@ export const useAppStore = defineStore('app', () => {
   const splitpaneResize = ref(0)
   const fromRoute = ref(null)
   const routerError = ref(null)
-  const luigiContext = ref(null)
 
   const { notify } = useNotification()
-
-  const isInIframe = computed(() => window.self !== window.top)
-
-  if (isInIframe.value) {
-    logger.debug('Registering listener for Luigi context initialization and context updates')
-    LuigiClient.addInitListener(context => setLuigiContext(context))
-    LuigiClient.addContextUpdateListener(context => setLuigiContext(context))
-  }
-
-  function setLuigiContext (value) {
-    luigiContext.value = value
-  }
-
-  const accountId = computed(() => luigiContext.value?.accountId)
-
-  function getLuigiContext () {
-    if (!isInIframe.value) {
-      return Promise.resolve(null)
-    }
-    if (luigiContext.value !== null) {
-      return Promise.resolve(luigiContext.value)
-    }
-    return new Promise(resolve => {
-      const timeout = 3000
-      const timeoutId = setTimeout(() => {
-        unwatch()
-        logger.error('The initialization of the Luigi Client has timed out after %d milliseconds', timeout)
-        resolve(null)
-      }, timeout)
-      const unwatch = watch(luigiContext, context => {
-        if (context !== null) {
-          clearTimeout(timeoutId)
-          unwatch()
-          resolve(context)
-        }
-      }, {
-        immediate: true,
-      })
-    })
-  }
-
-  function setRoute (route) {
-    if (isInIframe.value) {
-      const pathname = toRef(route, 'path')
-      watch(pathname, value => {
-        if (value) {
-          LuigiClient.linkManager().fromVirtualTreeRoot().withoutSync().navigate(value)
-        }
-      }, {
-        immediate: true,
-      })
-    }
-  }
 
   function updateSplitpaneResize () {
     splitpaneResize.value = Date.now()
@@ -149,12 +85,7 @@ export const useAppStore = defineStore('app', () => {
     focusedElementId,
     splitpaneResize,
     fromRoute,
-    isInIframe,
-    setRoute,
     routerError,
-    accountId,
-    luigiContext,
-    getLuigiContext,
     updateSplitpaneResize,
     setError,
     setHeaderWarning,
