+++ frontend/src/views/GAdministration.vue
@@ -511,6 +511,7 @@ import GShootCustomFieldsConfiguration from '@/components/GShootCustomFieldsConf
 import GResourceQuotaHelp from '@/components/GResourceQuotaHelp.vue'
 import GTextRouterLink from '@/components/GTextRouterLink.vue'
 
+import { useLuigiContext } from '@/composables/useLuigiContext'
 import { useProvideProjectItem } from '@/composables/useProjectItem'
 import { useProvideProjectContext } from '@/composables/useProjectContext'
 import { useLogger } from '@/composables/useLogger'
@@ -528,6 +529,7 @@ import set from 'lodash/set'
 import includes from 'lodash/includes'
 
 const logger = useLogger()
+const luigiContext = useLuigiContext()
 const appStore = useAppStore()
 const configStore = useConfigStore()
 const quotaStore = useQuotaStore()
@@ -540,7 +542,10 @@ const kubeconfigStore = useKubeconfigStore()
 const route = useRoute()
 const router = useRouter()
 
-useProvideProjectContext()
+useProvideProjectContext({
+  luigiContext,
+  configStore,
+})
 
 const color = ref('primary')
 const errorMessage = ref(undefined)
@@ -646,9 +651,9 @@ async function updateProperty (key, value, options = {}) {
       metadata: { name },
       spec: { namespace },
     }
-    if (appStore.accountId && !get(projectStore.project, ['metadata', 'annotations', 'openmfp.org/account-id'])) {
+    if (luigiContext.accountId && !get(projectStore.project, ['metadata', 'annotations', 'openmfp.org/account-id'])) {
       set(mergePatchDocument, ['metadata', 'labels', 'openmfp.org/managed-by'], 'true')
-      set(mergePatchDocument, ['metadata', 'annotations', 'openmfp.org/account-id'], appStore.accountId)
+      set(mergePatchDocument, ['metadata', 'annotations', 'openmfp.org/account-id'], luigiContext.accountId)
     }
     set(mergePatchDocument, ['spec', key], value)
     await projectStore.patchProject(mergePatchDocument)
